requirements.txt
# Web Framework
apiflask==2.1.0
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Async Support
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# WSGI Server
gunicorn==21.2.0 ; sys_platform != 'win32'
waitress==2.1.2 ; sys_platform == 'win32'

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-flask==1.3.0
faker==20.1.0

# Code Quality
flake8==6.1.0
black==23.12.1
mypy==1.7.1

# CORS (Optional)
flask-cors==4.0.0

# Performance Testing (Optional)
locust==2.19.1

